backend/requirements.txt
Flask
Flask-CORS
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
openai
requests
gunicorn 